+++ core/src/main/java/tech/pegasys/web3signer/core/FilecoinRunner.java
@@ -41,6 +41,7 @@ import tech.pegasys.web3signer.core.signing.filecoin.FilecoinNetwork;
 import java.util.List;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.json.JsonMapper;
 import com.github.arteam.simplejsonrpc.server.JsonRpcServer;
 import io.vertx.core.Vertx;
 import io.vertx.ext.web.Router;
@@ -86,11 +87,12 @@ public class FilecoinRunner extends Runner {
     final FcJsonRpcMetrics fcJsonRpcMetrics = new FcJsonRpcMetrics(metricsSystem);
     final FcJsonRpc fileCoinJsonRpc = new FcJsonRpc(fcSigners, fcJsonRpcMetrics);
     final ObjectMapper mapper =
-        new ObjectMapper()
+        JsonMapper.builder()
             .enable(ACCEPT_CASE_INSENSITIVE_ENUMS)
             .enable(ACCEPT_CASE_INSENSITIVE_PROPERTIES)
             .disable(FAIL_ON_UNKNOWN_PROPERTIES)
-            .registerModule(new FilecoinJsonRpcModule());
+            .addModule(new FilecoinJsonRpcModule())
+            .build();
     final JsonRpcServer jsonRpcServer = JsonRpcServer.withMapper(mapper);
 
     router