+++ core/src/main/java/tech/pegasys/web3signer/core/service/http/SigningObjectMapperFactory.java
@@ -45,7 +45,9 @@ import tech.pegasys.web3signer.core.service.http.handlers.signing.eth2.json.Bloc
 import com.fasterxml.jackson.core.Version;
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.MapperFeature;
+import com.fasterxml.jackson.databind.Module;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.json.JsonMapper;
 import com.fasterxml.jackson.databind.module.SimpleModule;
 import org.apache.tuweni.bytes.Bytes;
 import org.apache.tuweni.bytes.Bytes32;
@@ -58,13 +60,14 @@ public class SigningObjectMapperFactory {
 
   private SigningObjectMapperFactory() {
     this.objectMapper =
-        new ObjectMapper()
+        JsonMapper.builder()
             .enable(MapperFeature.ACCEPT_CASE_INSENSITIVE_ENUMS)
-            .disable(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES);
-    addWeb3SignerMappers();
+            .disable(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES)
+            .addModule(web3SignerMappers())
+            .build();
   }
 
-  private void addWeb3SignerMappers() {
+  private Module web3SignerMappers() {
     final SimpleModule module =
         new SimpleModule("SigningJsonRpcModule", new Version(1, 0, 0, null, null, null));
     module.addDeserializer(Bytes.class, new HexDeserialiser());
@@ -95,9 +98,9 @@ public class SigningObjectMapperFactory {
     module.addDeserializer(SszBitvector.class, new SszBitvectorDeserializer());
     module.addSerializer(SszBitvector.class, new SszBitvectorSerializer());
 
-    module.addDeserializer(BlockRequest.class, new BlockRequestDeserializer(objectMapper));
+    module.addDeserializer(BlockRequest.class, new BlockRequestDeserializer());
 
-    objectMapper.registerModule(module);
+    return module;
   }
 
   public static ObjectMapper createObjectMapper() {